### 2019-03-14 #289 @laurajanegraham
Submission: grainchanger
Submitting Author: Name (@laurajanegraham)
Repository: https://github.com/laurajanegraham/grainchanger
Version submitted:  0.1.0 .9000
Version accepted: TBD

---

```
Package: grainchanger
Title: Moving-Window and Direct Data Aggregation
Version: 0.1.0.9000
Authors@R: c(person("Laura", "Graham",
             email = "LauraJaneEGraham@gmail.com",
             person("Felix", "Eigenbrod",
             role = "ctb",
             email = "f.eigenbrod@soton.ac.uk",
             comment = "Input on initial conceptual development"),
             person("Marco", "Sciaini",
             role = "ctb",
             email = "sciaini.marco@gmail.com",
             comment = "Input on package development and structure"))
Description: Data aggregation via moving window or direct methods. Aggregate a
    fine-resolution raster to a grid. The moving window method smooths the surface
    using a specified function within a moving window of a specified size and shape
    prior to aggregation. The direct method simply aggregates to the grid using the
    specified function.
    raster,
    sf,
    furrr,
    checkmate,
    methods
    covr,
    landscapetools
Language: en-GB

```

The package provides utilities to aggregate fine resolution spatial data to a coarser resolution grid through 2 methods: (1) moving-window data aggregation; and (2) direct data aggregation.

Spatial scientists (geographers, spatial ecologists, social scientists, anyone who works with spatial data)

No. It is possible to do these analyses in the raster package, but it certainly isn't trivial to do so. We have streamlined the process for (a) ease of use; and (b) computational speed.
