### 2019-03-28 5065a04
added in line to make prop function work for nomove
