### 2019-04-10 #289 @mbjoseph
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

Really neat package @laurajanegraham! Overall I think this is in great shape.

- A description in the README that outlines the similarities and differences to `raster::focal()`, and how the moving window functionality relates to spatial convolution would be helpful. It took me a while to wrap my head around what the package did based on the README, but a more explicit connection to other operations that users may know already could reduce the spin-up time for new users.

- It could be good to point out the target audience more directly in the README - i.e., some explicit statement that a user could identify with - landscape ecologists, or generally folks working with spatial data at different resolutions, etc.

- The three panel figure in the README (mwda_schematic.png) I thought would be helpful to wrap my head around what's happening, but there were a few things that I found confusing:

1. The terms "response" and "predictor" along with f(x) in the first panel might suggest some kind of model, but these aren't explained elsewhere.

2. It would be nice to have an explanation for why the number of grid cells increases in the second figure. Was a torus created?

3. An explanation of what the thick black squares represent would be useful.

- The `show_landscape` function applied to a categorical raster might be improved by using a discrete colormap (via `discrete = TRUE`). I was confused briefly by the current example, where the colorbar implies continuity.

- In the README there seem to be some places where "surface" and "landscape" are used interchangeably. Would using one or the other, (or "raster") generalize the language a bit?

#### Implementation and interface

- Some of the function arguments are very concise, possibly at the expense of clarity. For example, the parameter `g` represents a grid, and the parameter `dat` represents a raster. Some more expressive argument names might be useful here (e.g., maybe `grd` and `ras`). This is definitely tricky because you might not want to use `grid` and `raster`, but something to think about.

- The `na.rm` argument to `nm_prop` appears to be unused - is that intentional?

- An example of how to use a custom function for moving window and direct aggregation might be useful for users who want to extend or modify the current functionality.

- In the help file for `winmove_agg`, it might be useful to elaborate on the description of the `type` argument to list the valid values (I can tell that these are at least "circle", "Gauss", and "rectangle", but maybe there are more?).

- Great to have the code of conduct, but it would be even better to also have contributing guidelines (either in the README or in a separate file) to outline your preferred way of having others contribute to the package.

- You could add a CodeMeta file for the package as described in the rOpenSci dev guide: https://ropensci.github.io/dev_guide/building.html#creating-metadata-for-your-package

- Missing a closing parenthesis on the Description field for `var_range` in the table of functions in the getting started vignette: https://laurajanegraham.github.io/grainchanger/articles/getting-started.html#functions

- Building out some unit tests for `winmove(...fun = 'prop')`, `winmove()` with user defined functions, and the uncovered functions in `R/calc_functions.R` would be a nice addition.

- For the `winmove` tests, it might be a good idea to use some different values of `d` - currently all of the tests use `d = 5`, which is great for that case, but as a consequence the tests might not reveal whether future changes break the functionality that enable different values of `d` to alter the function's behavior.

- If possible, it would be nice to have testing on Windows via AppVeyor.

#### Generic checks

- `goodpractice` checks are pretty clean, but do generate some potentially useful suggestions:

```r
> goodpractice::gp()
...

    R/calc_functions.R:11:NA
    R/calc_functions.R:12:NA
    R/calc_functions.R:43:NA
    R/calc_functions.R:44:NA
    R/calc_functions.R:47:NA

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80 characters

    R/create_torus.R:41:1
    R/create_torus.R:43:1
    R/create_torus.R:47:1
    R/create_torus.R:49:1
    R/create_torus.R:51:1

  ✖ avoid sapply(), it is not type safe. It might return a vector, or a list, depending on the input data. Consider using

    R/calc_functions.R:34:8

```

- A quick spell check raises a bunch of false positives as usual, but also some valid misspellings like "Ecololgy":

```r
> spelling::spell_check_package()
  WORD                       FOUND IN
Codecov                    README.md:11
                           README.Rmd:20
CRS                        winmove_agg.Rd:15
                           winmove.Rd:12
doi                        cat_ls.Rd:11
                           cont_ls.Rd:11
                           g_sf.Rd:11
Ecololgy                   cat_ls.Rd:10
                           cont_ls.Rd:10
                           g_sf.Rd:10
Fritsch                    cat_ls.Rd:9,13
                           cont_ls.Rd:9
                           g_sf.Rd:9
grainchanger’            README.md:161
landscapetools             cat_ls.Rd:9,13,14
                           cont_ls.Rd:9
                           g_sf.Rd:9
lc                         README.md:128,130,132,133
NLMR                       cat_ls.Rd:9
                           cont_ls.Rd:9
                           g_sf.Rd:9
                           README.md:144
                           README.Rmd:119
                           getting-started.Rmd:100
org                        cat_ls.Rd:11,14
                           cont_ls.Rd:11
                           g_sf.Rd:11
rm                         nomove_agg.Rd:17
                           winmove_agg.Rd:24
                           winmove.Rd:21
shei                       README.md:130,132
Simpkins                   cat_ls.Rd:9
                           cont_ls.Rd:9
                           g_sf.Rd:9
sp                         nomove_agg.Rd:23
SpatialPolygonsDataFrame   nomove_agg.Rd:11
                           winmove_agg.Rd:11
useage                     README.md:118
                           README.Rmd:91
                           getting-started.Rmd:72
var                        README.md:134
wm                         README.md:128,129,130,131
```
