### 2019-04-12 #289 @laurajanegraham
Thanks for your thorough review @mbjoseph - that's all really useful.

I'll deal with the comments all at once when we have the second review through, but in the meantime @karthik how would you recommend going about changing parameter names for a package that may already be in use? I agree that they should be made more meaningful.
