### 2019-04-17 1ed8d90
update so that can take in non-grid polygon as larger resolution
