+++ R/nomove_agg.R
@@ -3,40 +3,41 @@
 #' Calculate the value for a given function for each cell in a larger resolution grid.
 #'
 #' @param g the grid across which to calculate the aggregated moving window function
-#'  (raster, SpatialPolygonsDataFrame, or sf object)
+#'   (raster, SpatialPolygonsDataFrame, or sf object)
 #' @param dat The raster dataset to aggregate
 #' @param fun The function to apply. The function fun should take multiple numbers, and
-#'  return a single number. For example mean, modal, min or max. It should also accept a
-#'  na.rm argument (or ignore it, e.g. as one of the 'dots' arguments. For example, length
-#'  will fail, but function(x, ...){na.omit(length(x))} works. See Details
+#'   return a single number. For example mean, modal, min or max. It should also accept a
+#'   na.rm argument (or ignore it, e.g. as one of the 'dots' arguments. For example,
+#'   length will fail, but function(x, ...){na.omit(length(x))} works. See Details
+#' @param is_grid Use \code{TRUE} if \code{g} contains only rectangular cells (i.e. a
+#'   grid). If \code{g} is any other polygon file, this should be set to false
 #' @param ... further arguments passed to or from other methods
 #'
 #' @return Raster (if input is Raster) or numeric vector (if input is sp or sf object)
-#'  containing values calculated for each coarser cell
+#'   containing values calculated for each coarser cell
 #'
 #' @keywords spatial, aggregate
 #'
-#' @details \code{grainchanger} has several built-in functions. Functions currently included are:
-#' \itemize{
-#' \item \code{nm_shei} - Shannon evenness, requires the additional argument \code{lc_class} (vector or scalar)
-#' \item \code{nm_prop} - Proportion, requires the additional argument \code{lc_class} (scalar)
-#' \item \code{var_range} - Range (max - min)
-#' }
+#' @details \code{grainchanger} has several built-in functions. Functions currently
+#'   included are: \itemize{ \item \code{nm_shei} - Shannon evenness, requires the
+#'   additional argument \code{lc_class} (vector or scalar) \item \code{nm_prop} -
+#'   Proportion, requires the additional argument \code{lc_class} (scalar) \item
+#'   \code{var_range} - Range (max - min) }
 #'
 #' @examples
 #' # load required data
 #' data(g_sf)
 #' data(cont_ls)
 #' data(cat_ls)
-#' 
+#'
 #' # aggregate using mean
 #' d <- nomove_agg(g_sf, cont_ls, "mean")
-#' 
+#'
 #' # aggregate using Shannon evenness
 #' d <- nomove_agg(g_sf, cont_ls, "nm_shei", lc_class = 0:3)
 #' @export
 
-nomove_agg <- function(g, dat, fun, ...) {
+nomove_agg <- function(g, dat, fun, is_grid = TRUE, ...) {
   checkmate::assert(
     checkmate::check_class(g, "RasterLayer"),
     checkmate::check_class(g, "SpatialPolygonsDataFrame"),
@@ -44,6 +45,8 @@ nomove_agg <- function(g, dat, fun, ...) {
   )
 
   checkmate::assert_class(dat, "RasterLayer")
+  
+  if(is_grid) warning("aggregation assumes all cells are rectangular\nset is.grid = FALSE if g is not a grid")
 
   if (fun == "shei") fun <- "nm_shei"
   if (fun == "prop") fun <- "nm_prop"
@@ -65,7 +68,11 @@ nomove_agg <- function(g, dat, fun, ...) {
     out <- furrr::future_map_dbl(sf::st_geometry(g), function(grid_cell, dat, fun, ...) {
[…]
