+++ R/winmove_agg.R
@@ -15,6 +15,8 @@
 #'  return a single number. For example mean, modal, min or max. It should also accept a
 #'  na.rm argument (or ignore it, e.g. as one of the 'dots' arguments. For example, length
 #'  will fail, but function(x, ...){na.omit(length(x))} works. See Details
+#' @param is_grid Use \code{TRUE} if \code{g} contains only rectangular cells (i.e. a
+#'   grid). If \code{g} is any other polygon file, this should be set to false
 #' @param ... further arguments passed to or from other methods
 #'
 #' @return Numeric vector containing moving window values calculated for each grid cell
@@ -41,7 +43,7 @@
 #' d <- winmove_agg(g_sf, cat_ls, 5, "rectangle", "shei", lc_class = 0:3)
 #' @export
 
-winmove_agg <- function(g, dat, d, type, fun, ...) {
+winmove_agg <- function(g, dat, d, type, fun, is_grid = TRUE, ...) {
   checkmate::assert(
     checkmate::check_class(g, "RasterLayer"),
     checkmate::check_class(g, "SpatialPolygonsDataFrame"),
@@ -51,6 +53,8 @@ winmove_agg <- function(g, dat, d, type, fun, ...) {
   checkmate::assert_class(dat, "RasterLayer")
   checkmate::assert_numeric(d)
 
+  if(is_grid) warning("aggregation assumes all cells are rectangular\nset is.grid = FALSE if g is not a grid")
+  
   # convert raster to grid
   if ("RasterLayer" %in% class(g)) {
     g <- methods::as(g, "SpatialPolygonsDataFrame")
@@ -64,7 +68,11 @@ winmove_agg <- function(g, dat, d, type, fun, ...) {
     grid_buffer <- sf::st_buffer(grid_cell, dist = d, endCapStyle = "SQUARE", joinStyle = "MITRE", mitreLimit = d / 2)
     grid_buffer <- sf::st_geometry(grid_buffer)
     grid_buffer <- sf::st_sf(grid_buffer)
-    dat_cell <- raster::crop(dat, grid_buffer)
+    if(is.grid) {
+      dat_cell <- raster::crop(dat, grid_buffer)  
+    } else {
+      dat_cell <- raster::mask(dat, grid_buffer)   # mask is slower, but needs to be used if polygons are not rectangular
+    }
     winmove_cellr <- winmove(dat_cell, d, type, fun, ...)
     mean(raster::values(winmove_cellr), na.rm = TRUE)
   }, dat, d, type, fun, ...)
