### 2019-04-18 83fefc8
amended so that raster is cropped before masking (quicker)
