### 2019-04-18 00bd65a
update so mask gets cropped version
