### 2019-04-25 #289 @karthik
Hi @laurajanegraham Thanks for the question. We have a section in the dev guide that addresses this specific issue: https://ropensci.github.io/dev_guide/evolution.html#functions-changing-function-names

I hope that's helpful. If not please let us know and we can help more.
