### 2019-04-30 #289 @johnbaums
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

Great submission, @laurajanegraham! I'm always on the lookout for packages that make spatial data munging more efficient - this is a nice one to add to the toolkit.

I've had a quick look through @mbjoseph's review and I second his suggestions regarding additional unit tests.

- I initially had some trouble getting the key functions to run. This was first due to an outdated version of dplyr (0.7.8; `error: object 'group_map' not found` ... dplyr 0.8.0.1 rectified this), and then due to an outdated version of R (3.4.2; `invalid class "Extent" object: TRUE` ... all tests ran succeeded after updating to R 3.5.3). It might be worth stipulating minimum versions for these in DESCRIPTION. (Note that the old R issue is also reported [here](https://cran.r-project.org/web/checks/check_results_grainchanger.html).)

- Nice but a little brief. It might be worth stating that `winmove` can be used on its own, without subsequently aggregating. I'd also like to see more examples, e.g. using rasters as the target grids. It would also benefit from a clearer description of the process involved in moving window aggregation. In particular, it's not clear that what is returned is the *average* of smoothed pixel values within the larger extents.
- I was a bit confused by the [schematic](https://github.com/laurajanegraham/grainchanger/blob/master/man/figures/mwda_schematic.png), until I realised that the leftmost grid represents one of the 16 grid cells shown in the middle grid.. I'm not sure how that could be more effectively portrayed.
- For people unfamiliar with construction of `sf` polygon grids, it'd be useful to include an example in the readme (and in the function docs) that creates `g_sf`.

- typos: Ecololgy, useage
- I'd like to see more details about the available functions (i.e. things in calc_functions.R). The names given in the Details section (e.g., 'wm_shei') throw errors, and it's not immediately clear that the wm_ prefix should be dropped by the user.
- It's unclear how you intend for `wm_classes` to be used with `winmove`.
- Might be an idea to provide a reference or formula for Shannon's evenness.

#### create_torus()
- I think that the output extent should be bsaed on the input raster. Currently, the extent is reset to `c(0, ncol(data_pad) * resolution, 0, nrow(dat_pad) * resolution)`. It would make more sense to use `extent(dat) + 2*r` so that this propagates nicely to other functions in the smoothing/aggregation pipeline (and also because `create_torus` will be useful independent of the pipeline).

- **Documentation:**
  - `r`: "The radius by which to pad the raster" - this should stipulate the units (i.e., number of cells). Actually I don't think "radius" is the best word choice here.
  - Define torus e.g. with "wrapping"/"tiling" - it's unclear without digging into the source, and the torus approach to padding may not always be appropriate so it should be made explicit. I had initially expected that "Pad a raster by a specified radius" meant it would be padded with NA, as per `raster::extend`). Though maybe the approach you're using is also what `raster::focal` does (where it's also not clearly described)?

#### winmove_agg()
- is `grid_buffer <- sf::st_geometry(grid_buffer)` necessary?
- I'm not sure whether this potentially a problem or not.. the distance, `d`, is used to define the moving window, but also to buffer the extent that is being considered (i.e. to buffer `grid_cell`). The latter is necessary to prevent edge effects of blocked moving window analysis, but I think the resulting raster (`winmove_cellr`) should then be cropped back to the extent of `grid_cell` prior to calculating the mean over the cells. Otherwise, raster cells `2d` away from the grid cell might(?) contribute to the aggregate mean of that grid cell. I think this may in fact not happen, since the cells at the margins are `NA` after `winmove`, at least in one test I ran.
- It's probably worth letting users define the function used for aggregation of smoothed cell values, and not just the moving window summary. Currently the function enforces calculation of the cells' mean. In contrast, the `nomove_agg` function uses the user-specified function for aggregation. This inconsistency might be confusing, and since it's trivial to allow users to specify the aggregation function as well, I think it's worth doing.
- `g` cannot be a `SpatialPolygons` object - if there's no harm in allowing that, I would do so (`checkmate::check_class(g, "SpatialPolygons")`).
- `winmove_agg` fails if `g` is a RasterLayer filled with `NA`, e.g.:
```
g <- raster(matrix(NA, ncol=2, nrow=2), xmn=0, xmx=8, ymn=0, ymx=8)
r <- raster(matrix(1:64, ncol=8), xmn=0, xmx=8, ymn=0, ymx=8)
winmove_agg(g, r, 2, 'rectangle', 'mean')
## Error in UseMethod("st_geometry") :
##  no applicable method for 'st_geometry' applied to an object of class "NULL"
```
- Should `winmove_agg` return a raster object by default if the target grid is a raster object? It currently returns a numeric vector. This is inconsistent with `nomove_agg`, which returns a raster if input is a raster.
- If the destination grid is a regular grid aligned with the initial raster (though alignment may be unecessary since `raster::crop` snaps to nearest cell by default), then it's much faster in some cases to use a `raster::focal` and `raster::aggregate` workflow. It might be worth testing for this and using that workflow where appropriate. For example:

```
# small raster:
r <- raster(matrix(runif(200^2), 200), xmn=0, xmx=200, ymn=0, ymx=200)
g <- raster(matrix(1, nc=40, nr=40), xmn=0, xmx=200, ymn=0, ymx=200)
system.time(d1 <- winmove_agg(g, r, 5, "rectangle", "mean"))
##   user  system elapsed
##  26.97    0.01   27.16

system.time({
  rf <- focal(r, raster::focalWeight(r, 5, 'rectangle'))
  d2 <- raster::aggregate(rf, 5, mean)
})
##   user  system elapsed
##   0.02    0.00    0.02

range(d1-d2[], na.rm=T)
## [1] -2.220446e-16  2.220446e-16

# scales well:
system.time({
  rf <- focal(r, raster::focalWeight(r, 5, 'rectangle'))
  d2 <- raster::aggregate(rf, 5, mean)
})
##   user  system elapsed
##   1.54    0.21    1.75
```
- It would be nice if these functions worked on raster stacks as well.

- **Documentation:**
    - docs list only a subset of the functions that exist in `calc_functions.R` ("grainchanger has several built-in functions. Functions currently included are: wm_shei, wm_prop, wm_classes, var_range"). This is missing `wm_mean`. I would also suggest not referring to the `nm_` and `wm_` prefixes in the docs. Users should be able to (and can) just use, e.g. "shei". I don't think they need to know the underlying function name, unless you intend for it to be used directly (which I doubt since it's not exported).
    - Function names could be more intuitive, e.g. "shannon" as opposed to "shei".
    -  `d`: "the size of the matrix returned" - I know that this relates to `raster::focalWeights`, but users might be confused about which matrix you refer to.
    - `fun`: use `\code{}` to mark up code.

#### winmove()
- is `grid_buffer <- sf::st_geometry(grid_buffer)` necessary?
-  `w <- ifelse(w > 0, 1, NA)` in `winmove` can break expected behaviour when `type='Gauss'`, for example if `fun=max`. E.g., compare `winmove(cont_ls, 5, 'Gauss', 'max')` to `raster::focal(cont_ls, raster::focalWeight(cont_ls, 5, 'Gauss'), max)`.
- I think it would be nice to accomodate anonymous functions in `winmove`. E.g.: `winmove(cont_ls, 5, 'rectangle', function(x, ...) diff(range(x)))`. This applies to `nomove_agg` as well.

- **Documentation:**
    - "The shape of the moving window" is unclear. What options can this take? Specify in Details or in the argument description. See `?raster::focalWeight`, for example.

#### nomove_agg()
- `nomove_agg` uses `as.vector()` while `winmove_agg` uses `raster::values()`. The latter is perhaps slightly more performant, though the difference is trivial. But you could change for consistency.
- when `g` is a raster layer, it's only used to determine the aggregation factor. However, users might expect that if the layers are offset, the aggregation will be done within the offset, target cells. This isn't the case.
- In `nm_shei`, I believe `p <- sum(dat == i) / raster::ncell(dat)` needs a `na.rm=TRUE`, and since `dat` is not a raster object in that scope, you could replace the `ncell` call with `length`. If you want to ignore NA cells from the number of cells (I'm not sure which is more appropriate), then you can replace the whole line with `p <- mean(dat==i, na.rm=TRUE)`.
- In `var_range`, can replace `if (sum(is.na(dat)) == length(dat))` with `if (all(is.na(dat)))`

- **Documentation:**
  - Use `\code{}` in docs (`fun` arg description)
  - I think the final example in `nomove_agg` (`d <- nomove_agg(g_sf, cont_ls, "nm_shei", lc_class = 0:3)`) should use `cat_ls` not `cont_ls`.

#### Tests and checks
- `devtools::test()` - all passed
- `devtools::check()` - no errors, warnings, or notes

#### sessionInfo
```
## R version 3.5.3 (2019-03-11)
## Platform: x86_64-w64-mingw32/x64 (64-bit)
## Running under: Windows >= 8 x64 (build 9200)
##
## Matrix products: default
##
## locale:
## [1] LC_COLLATE=English_Australia.1252  LC_CTYPE=English_Australia.1252    LC_MONETARY=English_Australia.1252 LC_NUMERIC=C
## [5] LC_TIME=English_Australia.1252
##
## attached base packages:
## [1] stats     graphics  grDevices utils     datasets  methods   base
##
## other attached packages:
## [1] purrr_0.3.2             grainchanger_0.1.0.9000 testthat_2.1.1          raster_2.8-19           sp_1.3-1
##
## loaded via a namespace (and not attached):
##  [1] Rcpp_1.0.1         compiler_3.5.3     class_7.3-15       prettyunits_1.0.2  remotes_2.0.4      tools_3.5.3        digest_0.6.18      pkgbuild_1.0.3
##  [9] pkgload_1.0.2      checkmate_1.9.1    memoise_1.1.0      lattice_0.20-38    rlang_0.3.4        DBI_1.0.0          cli_1.1.0          rstudioapi_0.10
## [17] commonmark_1.7     parallel_3.5.3     yaml_2.2.0         xopen_1.0.0        e1071_1.7-1        furrr_0.1.0        withr_2.1.2        stringr_1.4.0
## [25] roxygen2_6.1.1     xml2_1.2.0         globals_0.12.4     desc_1.2.0         fs_1.2.7           devtools_2.0.2     classInt_0.3-3     rprojroot_1.3-2
## [33] grid_3.5.3         glue_1.3.1         listenv_0.7.0      sf_0.7-4           R6_2.4.0           processx_3.3.0     rcmdcheck_1.3.2    sessioninfo_1.1.1
## [41] callr_3.2.0        magrittr_1.5       units_0.6-2        backports_1.1.4    codetools_0.2-16   ps_1.3.0           usethis_1.5.0      assertthat_0.2.1
## [49] future_1.12.0      KernSmooth_2.23-15 stringi_1.4.3      crayon_1.3.4
```
