### 2019-05-07 c4d4e42
allow users to choose the aggregation function
