### 2019-05-09 6495b60
update wm_prop so that it returns 0 if the lc_class is not there
