### 2019-05-09 a24c7fd
deal with NAs in the conditional
