+++ R/nomove_agg.R
@@ -2,15 +2,19 @@
 #'
 #' Calculate the value for a given function for each cell in a larger resolution grid.
 #'
-#' @param g the grid across which to calculate the aggregated moving window function
-#'   (raster, SpatialPolygonsDataFrame, or sf object)
-#' @param dat The raster dataset to aggregate
-#' @param fun The function to apply. The function fun should take multiple numbers, and
+#' @param coarse_dat  sf, Raster* or Spatial* object. The coarse grain data (response
+#'   data) across which to calculate the aggregated function
+#' @param fine_dat Raster* object. Raster* object. The fine grain data (predictor /
+#'   covariate data) to aggregate
+#' @param fun character. The function to apply. The function fun should take multiple numbers, and
 #'   return a single number. For example mean, modal, min or max. It should also accept a
 #'   na.rm argument (or ignore it, e.g. as one of the 'dots' arguments. For example,
 #'   length will fail, but function(x, ...){na.omit(length(x))} works. See Details
-#' @param is_grid Use \code{TRUE} if \code{g} contains only rectangular cells (i.e. a
-#'   grid). If \code{g} is any other polygon file, this should be set to false
+#'  @param is_grid logical. Use \code{TRUE} (default) if \code{g} contains only rectangular
+#'   cells (i.e. a grid). If \code{g} is any other polygon file, this should be set to
+#'   false
+#' @param quiet logical. If \code{FALSE} (default) and \code{is_grid == TRUE} the user
+#'   gets a warning that the aggregation assumes all cells are rectangular
 #' @param ... further arguments passed to or from other methods
 #'
 #' @return Raster (if input is Raster) or numeric vector (if input is sp or sf object)
@@ -19,10 +23,12 @@
 #' @keywords spatial, aggregate
 #'
 #' @details \code{grainchanger} has several built-in functions. Functions currently
-#'   included are: \itemize{ \item \code{nm_shei} - Shannon evenness, requires the
-#'   additional argument \code{lc_class} (vector or scalar) \item \code{nm_prop} -
-#'   Proportion, requires the additional argument \code{lc_class} (scalar) \item
-#'   \code{var_range} - Range (max - min) }
+#'   included are: \itemize{ 
+#'      \item \code{shei} - Shannon evenness, requires the additional argument \code{lc_class} (vector or scalar) 
+#'      \item \code{prop} - Proportion, requires the additional argument \code{lc_class} (scalar) 
+#'      \item \code{classes} - Unique number of classes in a categorical landscape 
+#'      \item \code{var_range} - Range (max - min) 
+#'   }
 #'
 #' @examples
 #' # load required data
@@ -37,35 +43,52 @@
 #' d <- nomove_agg(g_sf, cont_ls, "nm_shei", lc_class = 0:3)
 #' @export
 
-nomove_agg <- function(g, dat, fun, is_grid = TRUE, ...) {
+nomove_agg <- function(coarse_dat, 
+                       fine_dat, 
+                       fun, 
+                       is_grid = TRUE, 
+                       quiet = FALSE, 
+                       g, dat, ...) {
+  
+  # code to deal with old parameter names
+  if (!missing(g)) {
+    warning("use 'coarse_dat' instead of 'g'\n")
+    coarse_dat <- g
+  }
+  
+  if (!missing(dat)) {
+    warning("use 'fine_dat' instead of 'dat'\n")
[…]
