+++ R/winmove_agg.R
@@ -1,80 +1,113 @@
 #' Moving-window data aggregation
 #'
-#' Calculate the mean moving window value for a given radius, shape and function for each cell
-#' in a larger resolution grid.
+#' Calculate the mean moving window value for a given radius, shape and function for each
+#' cell in a larger resolution grid.
 #'
-#' @param g the grid across which to calculate the aggregated moving window function
-#'  (raster, SpatialPolygonsDataFrame, or sf object)
-#' @param dat The raster dataset to aggregate
-#' @param d numeric. If \code{type=circle}, the radius of the circle (in units of the CRS). If
-#'  \code{type=rectangle} the dimension of the rectangle (one or two numbers). If \code{type=Gauss}
-#'  the size of sigma, and optionally another number to determine the size of the matrix
-#'  returned (default is 3 times sigma)
-#' @param type The shape of the moving window
-#' @param fun The function to apply. The function fun should take multiple numbers, and
-#'  return a single number. For example mean, modal, min or max. It should also accept a
-#'  na.rm argument (or ignore it, e.g. as one of the 'dots' arguments. For example, length
-#'  will fail, but function(x, ...){na.omit(length(x))} works. See Details
-#' @param is_grid Use \code{TRUE} if \code{g} contains only rectangular cells (i.e. a
-#'   grid). If \code{g} is any other polygon file, this should be set to false
+#' @param coarse_dat sf, Raster* or Spatial* object. The coarse grain data (response data)
+#'   across which to calculate the aggregated moving window function
+#' @param fine_dat Raster* object. The fine grain data (predictor / covariate data) to
+#'   aggregate
+#' @param d numeric. If \code{type=circle}, the radius of the circle (in units of the
+#'   CRS). If \code{type=rectangle} the dimension of the rectangle (one or two numbers).
+#'   If \code{type=Gauss} the size of sigma, and optionally another number to determine
+#'   the size of the matrix returned (default is 3 times sigma)
+#' @param type character. The shape of the moving window
+#' @param win_fun character. The function to apply to the moving window. The function fun
+#'   should take multiple numbers, and return a single number. For example mean, modal,
+#'   min or max. It should also accept a na.rm argument (or ignore it, e.g. as one of the
+#'   'dots' arguments. For example, length will fail, but function(x,
+#'   ...){na.omit(length(x))} works. See Details
+#' @param agg_fun character. The function by which to aggregate. By default this is set to
+#'   \code{mean}
+#' @param is_grid logical. Use \code{TRUE} (default) if \code{g} contains only rectangular
+#'   cells (i.e. a grid). If \code{g} is any other polygon file, this should be set to
+#'   false
+#' @param quiet logical. If \code{FALSE} (default) and \code{is_grid == TRUE} the user
+#'   gets a warning that the aggregation assumes all cells are rectangular
 #' @param ... further arguments passed to or from other methods
 #'
 #' @return Numeric vector containing moving window values calculated for each grid cell
 #'
 #' @keywords focal, spatial, aggregate
 #'
-#' @details \code{grainchanger} has several built-in functions. Functions currently included are:
-#' \itemize{
[…]
