+++ DESCRIPTION
@@ -36,7 +36,7 @@ Suggests:
     covr,
     ggplot2,
     landscapetools
-RoxygenNote: 6.1.0
+RoxygenNote: 6.1.1
 Language: en-GB
 VignetteBuilder: knitr
 URL: https://github.com/laurajanegraham/grainchanger

[…]
