### 2019-05-14 b96d14a
update winmove code to not output cropped data
