### 2019-05-21 d3048f9
fixed input to nm_prop to remove na.rm arg
