### 2019-05-21 539696a
correcting minor typos
