+++ README.Rmd
@@ -16,22 +16,24 @@ knitr::opts_chunk$set(
 # grainchanger <img src="man/figures/logo.png" align="right" width="150" />
 
 <!-- badges: start -->
-  [![Travis build status](https://travis-ci.org/laurajanegraham/grainchanger.svg?branch=master)](https://travis-ci.org/laurajanegraham/grainchanger)
-  [![Codecov test coverage](https://codecov.io/gh/laurajanegraham/grainchanger/branch/master/graph/badge.svg)](https://codecov.io/gh/laurajanegraham/grainchanger?branch=master)
+[![Travis build status](https://travis-ci.org/laurajanegraham/grainchanger.svg?branch=master)](https://travis-ci.org/laurajanegraham/grainchanger)
+[![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/laurajanegraham/grainchanger?branch=master&svg=true)](https://ci.appveyor.com/project/laurajanegraham/grainchanger)
+[![Codecov test coverage](https://codecov.io/gh/laurajanegraham/grainchanger/branch/master/graph/badge.svg)](https://codecov.io/gh/laurajanegraham/grainchanger?branch=master)
 [![CRAN status](https://www.r-pkg.org/badges/version/grainchanger)](https://cran.r-project.org/package=grainchanger)
 [![](https://cranlogs.r-pkg.org/badges/grainchanger)](https://cran.r-project.org/package=grainchanger)
-[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+[![Project Status: Active â€“ The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+
 <!-- badges: end -->
 
 The `grainchanger` package provides functionality for data aggregation to a grid via moving-window or direct methods. 
 
 ### Moving-window data aggregation
 
-The moving-window data aggregation method smooths the surface using a specified function within a moving window of a specified size and shape prior to aggregation. 
+The moving-window data aggregation method smooths the raster using a specified function within a moving window of a specified size and shape prior to aggregation. 
 
 ![Schematic of the moving-window data aggregation approach](man/figures/mwda_schematic.png)
 
-The moving window approach allows users to capture some information about landscape structure at the scale at which the process acts in the landscape.
+The moving window approach allows users to capture some information about spatial structure at the scale at which the process acts on a particular ecological response.
 
 ### Direct data aggregation 
 
@@ -90,7 +92,7 @@ ggplot(g_sf) + geom_sf(aes(fill = dda))
 
 ## Functions
 
-There are a number of inbuilt functions in the grainchanger package, with their useage outlined below. While it is possible to use user-defined functions within both `winmove_agg` and `nomove_agg`, we welcome suggestions for additional functions. Please [add as an issue](https://github.com/laurajanegraham/grainchanger/issues) - doing it this way means we can maximise the speed of the function. 
+There are a number of inbuilt functions in the grainchanger package, with their usage outlined below. While it is possible to use user-defined functions within both `winmove_agg` and `nomove_agg`, we welcome suggestions for additional functions. Please [add as an issue](https://github.com/laurajanegraham/grainchanger/issues) - doing it this way means we can maximise the speed of the function. 
 
 ```{r functions, echo = FALSE}
 function_overview <- data.frame(
@@ -112,7 +114,7 @@ knitr::kable(function_overview)
 
 ### Create torus
 
[…]
