+++ README.md
@@ -7,14 +7,17 @@
 
 [![Travis build
 status](https://travis-ci.org/laurajanegraham/grainchanger.svg?branch=master)](https://travis-ci.org/laurajanegraham/grainchanger)
+[![AppVeyor build
+status](https://ci.appveyor.com/api/projects/status/github/laurajanegraham/grainchanger?branch=master&svg=true)](https://ci.appveyor.com/project/laurajanegraham/grainchanger)
 [![Codecov test
 coverage](https://codecov.io/gh/laurajanegraham/grainchanger/branch/master/graph/badge.svg)](https://codecov.io/gh/laurajanegraham/grainchanger?branch=master)
 [![CRAN
 status](https://www.r-pkg.org/badges/version/grainchanger)](https://cran.r-project.org/package=grainchanger)
 [![](https://cranlogs.r-pkg.org/badges/grainchanger)](https://cran.r-project.org/package=grainchanger)
-[![Project Status: Active – The project has reached a stable, usable
+[![Project Status: Active â€“ The project has reached a stable, usable
 state and is being actively
 developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+
 <!-- badges: end -->
 
 The `grainchanger` package provides functionality for data aggregation
@@ -22,7 +25,7 @@ to a grid via moving-window or direct methods.
 
 ### Moving-window data aggregation
 
-The moving-window data aggregation method smooths the surface using a
+The moving-window data aggregation method smooths the raster using a
 specified function within a moving window of a specified size and shape
 prior to aggregation.
 
@@ -30,8 +33,8 @@ prior to aggregation.
 approach](man/figures/mwda_schematic.png)
 
 The moving window approach allows users to capture some information
-about landscape structure at the scale at which the process acts in the
-landscape.
+about spatial structure at the scale at which the process acts on a
+particular ecological response.
 
 ### Direct data aggregation
 
@@ -114,6 +117,8 @@ show_landscape(cont_ls)
 g_sf$dda <- nomove_agg(coarse_dat = g_sf,
                        fine_dat = cont_ls, 
                        fun = "var_range")
+#> aggregation assumes all cells are rectangular
+#> <U+25CF> set `is_grid = FALSE` if coarse_dat is not a grid
 
 ggplot(g_sf) + geom_sf(aes(fill = dda))
 ```
@@ -123,7 +128,7 @@ ggplot(g_sf) + geom_sf(aes(fill = dda))
 ## Functions
 
 There are a number of inbuilt functions in the grainchanger package,
-with their useage outlined below. While it is possible to use
+with their usage outlined below. While it is possible to use
 user-defined functions within both `winmove_agg` and `nomove_agg`, we
 welcome suggestions for additional functions. Please [add as an
 issue](https://github.com/laurajanegraham/grainchanger/issues) - doing
@@ -141,10 +146,10 @@ function.
 
 ### Create torus
 
-The `create_torus` function takes as input a square or rectangular
-landscape and pads it by a specified radius, creating the effect of a
-torus. We developed this function in order to avoid edge effects when
-testing methods on simulated landscapes (such as those from
+The `create_torus` function takes as input a raster and pads it by a
+specified radius, creating the effect of a torus. We developed this
+function in order to avoid edge effects when testing methods on
+simulated rasters (such as those from
 [NLMR](https://ropensci.github.io/NLMR/)).
 
 ``` r
