### 2019-05-21 ec2bf86
re-document and fix winmove_agg example
