+++ R/winmove_agg.R
@@ -53,9 +53,9 @@
 winmove_agg <- function(coarse_dat, 
                         fine_dat, 
                         d, 
+                        type=c('circle', 'Gauss', 'rectangle'), 
                         win_fun, 
                         agg_fun = "mean", 
-                        type=c('circle', 'Gauss', 'rectangle'), 
                         is_grid = TRUE, 
                         quiet = FALSE, 
                         g, dat, fun,
