+++ README.Rmd
@@ -25,15 +25,19 @@ knitr::opts_chunk$set(
 
 <!-- badges: end -->
 
-The `grainchanger` package provides functionality for data aggregation to a grid via moving-window or direct methods. 
+The `grainchanger` package provides functionality for data aggregation to a coarser resolution via moving-window or direct methods. 
+
+As landscape ecologists and macroecologists, we often need to aggregate data in order to harmonise datasets. In doing so, we often lose a lot of information about the spatial structure and environmental heterogeneity of data measured at finer resolution. We developed this method and R package in order to overcome some of these issues. 
+
+For more information on the background to and motivation for the development of this method, see [Graham et al. 2019 in Methods in Ecology and Evolution](https://doi.org/10.1111/2041-210X.13177). 
 
 ### Moving-window data aggregation
 
-The moving-window data aggregation method smooths the raster using a specified function within a moving window of a specified size and shape prior to aggregation. 
+The moving-window data aggregation method smooths an input raster using a specified function within a moving window of a specified size and shape prior to aggregation. This acts as a convenient wrapper for the `focalWeight()` and `focal()` functions in the `raster` package. Additionally, we have aimed to write efficient functions for some oft-used metrics within landscape ecology for use within the moving window.
 
 ![Schematic of the moving-window data aggregation approach](man/figures/mwda_schematic.png)
 
-The moving window approach allows users to capture some information about spatial structure at the scale at which the process acts on a particular ecological response.
+
 
 ### Direct data aggregation 
 
