### 2019-05-21 7f18fc1
update readme for better use case and explanation of how fits with raster focal etc.
