### 2019-05-21 7927308
update torus to match extent of input raster (plus radius)
