### 2019-05-21 1a60ab5
update torus docs to explain the dpad (formally r) parameter
