### 2019-05-22 c73306d
fix winmove functions to deal with na properly in mean and to deal with data smaller than buffer
