+++ man/winmove.Rd
@@ -7,6 +7,8 @@
 winmove(fine_dat, d, type, win_fun, ...)
 }
 \arguments{
+\item{fine_dat}{The raster dataset on which to calculate the moving window function}
+
 \item{d}{numeric. If \code{type=circle}, the radius of the circle (in units of the CRS). If
 \code{type=rectangle} the dimension of the rectangle (one or two numbers). If \code{type=Gauss}
 the size of sigma, and optionally another number to determine the size of the matrix
@@ -14,14 +16,12 @@ returned (default is 3 times sigma)}
 
 \item{type}{The shape of the moving window}
 
-\item{...}{further arguments passed to or from other methods}
-
-\item{dat}{The raster dataset on which to calculate the moving window function}
-
-\item{fun}{The function to apply. The function fun should take multiple numbers, and
+\item{win_fun}{The function to apply. The function fun should take multiple numbers, and
 return a single number. For example mean, modal, min or max. It should also accept a
 na.rm argument (or ignore it, e.g. as one of the 'dots' arguments. For example, length
 will fail, but function(x, ...){na.omit(length(x))} works. See Details}
+
+\item{...}{further arguments passed to or from other methods}
 }
 \value{
 A smoothed raster with the moving window values calculated
