+++ man/winmove_agg.Rd
@@ -6,7 +6,7 @@
 \usage{
 winmove_agg(coarse_dat, fine_dat, d, type = c("circle", "Gauss",
   "rectangle"), win_fun, agg_fun = "mean", is_grid = TRUE,
-  quiet = FALSE, g, dat, fun, ...)
+  quiet = FALSE, ...)
 }
 \arguments{
 \item{coarse_dat}{sf, Raster* or Spatial* object. The coarse grain data (response data)
