### 2019-05-22 f28c231
update nomove with new parameter names
