+++ man/nomove_agg.Rd
@@ -4,8 +4,8 @@
 \alias{nomove_agg}
 \title{Direct data aggregation}
 \usage{
-nomove_agg(coarse_dat, fine_dat, fun, is_grid = TRUE, quiet = FALSE, g,
-  dat, ...)
+nomove_agg(coarse_dat, fine_dat, agg_fun, is_grid = TRUE,
+  quiet = FALSE, ...)
 }
 \arguments{
 \item{coarse_dat}{sf, Raster* or Spatial* object. The coarse grain data (response
@@ -14,7 +14,7 @@ data) across which to calculate the aggregated function}
 \item{fine_dat}{Raster* object. Raster* object. The fine grain data (predictor /
 covariate data) to aggregate}
 
-\item{fun}{character. The function to apply. The function fun should take multiple numbers, and
+\item{agg_fun}{character. The function to apply. The function fun should take multiple numbers, and
  return a single number. For example mean, modal, min or max. It should also accept a
  na.rm argument (or ignore it, e.g. as one of the 'dots' arguments. For example,
  length will fail, but function(x, ...){na.omit(length(x))} works. See Details
