### 2019-05-22 ba0c1ac
fix na handling
