+++ R/calc_functions.R
@@ -29,11 +29,7 @@ wm_shei <- function(dat, d, type, lc_class) {
 
 #' @noRd
 # mean (it's quicker using weighted sum than mean, but mean needed if na.rm required)
-wm_mean <- function(dat, d, type, ...) {
-  args <- list(...)
-  exist <- "na.rm" %in% names(args)
-  if(!exist) na.rm = FALSE
-  
+wm_mean <- function(dat, d, type, na.rm = FALSE) {
   if(na.rm) {
     w <- ifelse(raster::focalWeight(dat, d, type) == 0, 0, 1)
     return(raster::focal(x = dat, w = w, fun = mean, na.rm = TRUE))
