### 2019-05-22 be916d8
update tests for wider coverage and reflect changes
