+++ tests/testthat/test-create_torus.R
@@ -2,5 +2,8 @@ context("create_torus")
 
 test_that("torus output is correct", {
   d <- create_torus(cat_ls, 5)
-  expect_equal(d, torus)
+  expect_equal(d, torus_5)
+  
+  d <- create_torus(cont_ls, 20)
+  expect_equal(d, torus_20)
 })
