+++ tests/testthat/test-winmove.R
@@ -6,9 +6,15 @@ test_that("shei calculation is correct", {
 })
 
 test_that("mean calculation is correct", {
-  d <- winmove(cont_ls, 5, "rectangle", "mean")
+  d <- winmove(cont_ls, 2, "Gauss", "mean")
   expect_equal(d, wm_mean_dat)
 })
+
+test_that("mean na.rm = TRUE calculation is correct", {
+  d <- winmove(cont_ls, 15, "circle", "mean", na.rm = TRUE)
+  expect_equal(d, wm_mean_na_dat)
+})
+
 
 
 test_that("output is raster of same resolution as input", {
