### 2019-05-22 b57814b
update winmove to output raster if input is raster
