### 2019-05-22 d73f307
update vignettes and readme with new function definitions
