+++ README.md
@@ -194,6 +194,6 @@ show_landscape(torus, discrete = TRUE)
   - License: GPL3
   - Get citation information for `grainchanger` in R doing
     `citation(package = 'grainchanger')`
-  - Please note that the ‘grainchanger’ project is released with a
+  - Please note that the `grainchanger` project is released with a
     [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By contributing
     to this project, you agree to abide by its terms.
