+++ README.Rmd
@@ -131,4 +131,4 @@ show_landscape(torus, discrete = TRUE)
 * Please [report any issues or bugs](https://github.com/laurajanegraham/grainchanger/issues/new/).
 * License: GPL3
 * Get citation information for `grainchanger` in R doing `citation(package = 'grainchanger')`
-* Please note that the 'grainchanger' project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By contributing to this project, you agree to abide by its terms.
+* Please note that the `grainchanger` project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By contributing to this project, you agree to abide by its terms.
