+++ vignettes/getting-started.Rmd
@@ -37,11 +37,11 @@ The below example shows the moving-window data aggregation in action. It aggrega
 
 ```{r mwda_example, fig.show='hold'}
 g_sf$mwda <- winmove_agg(
-  g = g_sf,
-  dat = cat_ls, 
+  coarse_dat = g_sf,
+  fine_dat = cat_ls, 
   d = 5,
   type = "rectangle",
-  fun = "shei",
+  win_fun = "shei",
   lc_class = 0:3
 )
 
@@ -60,9 +60,9 @@ The below example shows the direct data aggregation in action. It aggregates a c
 show_landscape(cont_ls)
 
 g_sf$dda <- nomove_agg(
-  g = g_sf,
-  dat = cont_ls, 
-  fun = "var_range")
+  coarse_dat = g_sf,
+  fine_dat = cont_ls, 
+  agg_fun = "var_range")
 
 ggplot(g_sf) + geom_sf(aes(fill = dda))
 ```
