### 2019-05-24 4e17c9d
update calc functions to be object oriented
