+++ R/winmove.R
@@ -32,10 +32,10 @@
 #' data(cont_ls)
 #'
 #' # calculate the moving window mean
-#' d <- winmove(cont_ls, 5, "rectangle", "mean")
+#' d <- winmove(cont_ls, 5, "rectangle", mean)
 #'
 #' # calculate the moving window Shannon evenness
-#' d <- winmove(cat_ls, 5, "rectangle", "shei", lc_class = 0:3)
+#' d <- winmove(cat_ls, 5, "rectangle", shei, lc_class = 0:3)
 #' @export
 
 winmove <- function(fine_dat, d, type, win_fun, ...) {
