+++ R/winmove_agg.R
@@ -44,10 +44,10 @@
 #' data(cat_ls)
 #'
 #' #' # aggregate using mean
-#' d <- winmove_agg(g_sf, cont_ls, 5, "rectangle", "mean")
+#' d <- winmove_agg(g_sf, cont_ls, 5, "rectangle", mean)
 #'
 #' # aggregate using Shannon evenness
-#' d <- winmove_agg(g_sf, cat_ls, 5, "rectangle", "shei", lc_class = 0:3)
+#' d <- winmove_agg(g_sf, cat_ls, 5, "rectangle", shei, lc_class = 0:3)
 #' @export
 
 winmove_agg <- function(coarse_dat, 
