+++ man/nomove_agg.Rd
@@ -50,10 +50,10 @@ data(cont_ls)
 data(cat_ls)
 
 # aggregate using mean
-d <- nomove_agg(g_sf, cont_ls, "mean")
+d <- nomove_agg(g_sf, cont_ls, mean)
 
 # aggregate using Shannon evenness
-d <- nomove_agg(g_sf, cont_ls, "nm_shei", lc_class = 0:3)
+d <- nomove_agg(g_sf, cont_ls, nm_shei, lc_class = 0:3)
 }
 \keyword{aggregate}
 \keyword{spatial,}
