### 2019-05-24 72e7dac
fix nomove_agg shei example
