+++ man/nomove_agg.Rd
@@ -53,7 +53,7 @@ data(cat_ls)
 d <- nomove_agg(g_sf, cont_ls, mean)
 
 # aggregate using Shannon evenness
-d <- nomove_agg(g_sf, cont_ls, nm_shei, lc_class = 0:3)
+d <- nomove_agg(g_sf, cont_ls, shei, lc_class = 0:3)
 }
 \keyword{aggregate}
 \keyword{spatial,}
