### 2019-05-24 5813940
add in non-grid coarse data and additional tests
