### 2019-05-24 b644ea2
update tests to 100% coverage
