### 2019-05-24 c60d9d7
updated to have better warnings when data does not extend to allow the moving window
