### 2019-05-24 3af0ca4
updated prop.winmove so that it only calculates if the specified value is in the data
