### 2019-05-26 69c55fa
update lc class numbers to match data
