### 2019-05-26 1337fb7
fix zero prop output
