### 2019-05-27 bc71c1a
update documentation to fully document all functions
