### 2019-05-27 4f0eb77
fix docs and errors from R CMD CHECK
