+++ R/data.R
@@ -1,4 +1,4 @@
-#' Example continuous raster
+#' Example continuous raster (fine_dat)
 #'
 #' An example map to show functionality on continuous surfaces.
 #'
@@ -10,7 +10,7 @@
 #'     00:1-9. https://doi.org/10.1111/2041-210X.13076
 "cont_ls"
 
-#' Example categorical raster
+#' Example categorical raster (fine_dat)
 #'
 #' An example map to show functionality on categorical surfaces.
 #'
@@ -24,7 +24,7 @@
 #'     https://CRAN.R-project.org/package=landscapetools
 "cat_ls"
 
-#' Example grid
+#' Example grid (coarse_dat)
 #'
 #' An example grid to show functionality when aggregating using an sf object.
 #'
@@ -35,3 +35,12 @@
 #'     for simulating and modifying neutral landscape models in R. Methods in Ecology and Evolution. 2018;
 #'     00:1-9. https://doi.org/10.1111/2041-210X.13076
 "g_sf"
+
+#' Example polygon (coarse_dat)
+#'
+#' An example non-gridded coarse data to show functionality when aggregating using an sf object.
+#'
+#' @details Generated with \code{sf::st_make_grid(sf::st_as_sfc(sf::st_bbox(cont_ls)), cellsize = 13, square = FALSE)}
+#'
+#' @format An sf object.
+"poly_sf"
