+++ R/calc_functions.R
@@ -1,6 +1,6 @@
 #' Arithmetic mean
 #'
-#' An extension to \code{mean} for objects of class \code{RasterWinmove}
+#' An extension to \code{mean} for objects of class \code{winmove}
 #'
 #' @param x RasterLayer. The data over which to calculate the mean value within a moving
 #'   window
@@ -13,7 +13,8 @@
 #'   computation proceeds. \code{na.rm = FALSE} (default) is optimised to be much faster
 #'   than \code{na.rm = TRUE}. The latter should only be used if \code{NA} values are
 #'   present
-#'
+#' @param ... further arguments passed to or from other methods
+#' 
 #' @return RasterLayer. A smoothed raster with the mean calculated within the specified
 #'   moving window
 #'   
@@ -24,8 +25,8 @@
 #' # load required data
 #' data(cont_ls)
 #' 
-#' # convert data to object of class RasterWinmove
-#' cont_ls <- new("RasterWinmove", cont_ls)
+#' # convert data to object of class winmove
+#' cont_ls <- new("winmove", cont_ls)
 #' 
 #' # aggregate using a circular window with radius 3
 #' d <- mean(cont_ls, d = 3, type = "circle")
@@ -34,7 +35,7 @@ mean <- function(x, ...) UseMethod("mean")
 
 #' @name mean
 #' @export
-mean.winmove <- function(x, d, type, na.rm = FALSE) {
+mean.winmove <- function(x, d, type, na.rm = FALSE, ...) {
   if(na.rm) {
     w <- ifelse(raster::focalWeight(x, d, type) == 0, 0, 1)
     return(raster::focal(x = x, w = w, fun = mean, na.rm = TRUE))
@@ -56,8 +57,9 @@ mean.winmove <- function(x, d, type, na.rm = FALSE) {
 #' @param type character. The shape of the moving window
 #' @param na.rm logical. indicates whether \code{NA} values should be stripped before the
 #'   computation proceeds. \code{na.rm = TRUE} is the default
+#' @param ... further arguments passed to or from other methods
 #'   
-#' @return If \code{class(x) == "RasterWinmove"}, a smoothed raster with the size of the range of values calculated within the specified
+#' @return If \code{class(x) == "winmove"}, a smoothed raster with the size of the range of values calculated within the specified
 #'   moving window
 #'   
 #'   If \code{class(x) == "numeric"}, a single value representing the size of the range of values in \code{x}
@@ -70,8 +72,8 @@ mean.winmove <- function(x, d, type, na.rm = FALSE) {
 #' data(cat_ls)
 #' data(cont_ls)
 #' 
-#' # convert data to object of class RasterWinmove
-#' cat_ls <- new("RasterWinmove", cat_ls)
+#' # convert data to object of class winmove
+#' cat_ls <- new("winmove", cat_ls)
 #' 
 #' # aggregate using a rectangular window with dimensions c(2,3)
 #' d <- range(cont_ls, d = c(2,3), type = "rectangle")
@@ -85,7 +87,7 @@ var_range <- function(x, ...) UseMethod("var_range")
 
 #' @name var_range
 #' @export
-var_range.winmove <- function(x, d, type, na.rm = TRUE) {
+var_range.winmove <- function(x, d, type, na.rm = TRUE, ...) {
   w <- ifelse(raster::focalWeight(x, d, type) == 0, 0, 1)
   return(raster::focal(x = x, w = w, fun = function(dat) {
     max(dat, na.rm = na.rm) - min(dat, na.rm = na.rm)
@@ -94,7 +96,7 @@ var_range.winmove <- function(x, d, type, na.rm = TRUE) {
 
 #' @name var_range
 #' @export
-var_range.numeric <- function(x, na.rm = TRUE) {
[…]
