+++ man/winmove-class.Rd
@@ -1,8 +1,8 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/winmove.R
 \docType{class}
-\name{RasterWinmove-class}
-\alias{RasterWinmove-class}
+\name{winmove-class}
+\alias{winmove-class}
 \title{An S4 class for use with winmove functions (extends RasterLayer)}
 \description{
 An S4 class for use with winmove functions (extends RasterLayer). Objects
@@ -30,7 +30,7 @@ Slots for RasterLayer and RasterBrick objects
 # load required data
 data(cat_ls)
 
-# set \\code{cat_ls} to object of class \\code{RasterWinmove}
-new("RasterWinmove", cat_ls)
+# set \\code{cat_ls} to object of class winmove
+new("winmove", cat_ls)
 
 }
