### 2019-05-31 9c98285
revamping documentation and finalising ropensci revisions
