+++ R/winmove_agg.R
@@ -10,15 +10,13 @@
 #'   data) to aggregate
 #' @param d numeric. If \code{type=circle}, the radius of the circle (in units
 #'   of the CRS). If \code{type=rectangle} the dimension of the rectangle (one
-#'   or two numbers). If \code{type=Gauss} the size of sigma, and optionally
-#'   another number to determine the size of the matrix returned (default is 3
-#'   times sigma)
+#'   or two numbers).
 #' @param type character. The shape of the moving window
 #' @param win_fun character. The function to apply to the moving window. The
-#'   function fun should take multiple numbers, and return a single number. For
-#'   example mean, modal, min or max. It should also accept a na.rm argument (or
-#'   ignore it, e.g. as one of the 'dots' arguments. For example, length will
-#'   fail, but function(x, ...){na.omit(length(x))} works. See Details
+#'   function \code{win_fun} should take multiple numbers, and return a single number. For
+#'   example \code{mean}, \code{modal}, \code{min} or \code{max}. It should also accept a \code{na.rm} argument (or
+#'   ignore it, e.g. as one of the 'dots' arguments. For example, \code{length} will
+#'   fail, but \code{function(x, ...){na.omit(length(x))}} works. See Details
 #' @param agg_fun character. The function by which to aggregate. By default this
 #'   is set to \code{mean}
 #' @param is_grid logical. Use \code{TRUE} (default) if \code{g} contains only
@@ -35,11 +33,15 @@
 #' @keywords focal, spatial, aggregate
 #'
 #' @details \code{grainchanger} has several built-in functions. Functions
-#'   currently included are: \itemize{ \item \code{shei} - Shannon evenness,
-#'   requires the additional argument \code{lc_class} (vector or scalar) \item
-#'   \code{prop} - Proportion, requires the additional argument \code{lc_class}
-#'   (scalar) \item \code{classes} - Unique number of classes in a categorical
-#'   landscape \item \code{var_range} - Range (max - min) }
+#'   currently included are: 
+#'   \itemize{ 
+#'      \item \code{shdi} - Shannon diversity, requires the additional argument \code{lc_class} (vector or scalar) 
+#'      \item \code{shei} - Shannon evenness, requires the additional argument \code{lc_class} (vector or scalar) 
+#'      \item \code{prop} - Proportion, requires the additional argument \code{lc_class} (scalar)
+#'      \item \code{var_range} - Range (max - min) 
+#'      }
+#'      
+#'  Note that \code{winmove_agg} can be run in parallel using \code{plan(multiprocess)} from the \code{future} package. 
 #' @examples
 #' # load required data
 #' data(g_sf)
@@ -56,7 +58,7 @@
 winmove_agg <- function(coarse_dat, 
                         fine_dat, 
                         d, 
-                        type=c('circle', 'Gauss', 'rectangle'), 
+                        type=c('circle', 'rectangle'), 
                         win_fun, 
                         agg_fun = mean, 
                         is_grid = TRUE, 
@@ -66,12 +68,14 @@ winmove_agg <- function(coarse_dat,
   checkmate::assert(
     checkmate::check_class(coarse_dat, "RasterLayer"),
     checkmate::check_class(coarse_dat, "SpatialPolygonsDataFrame"),
+    checkmate::check_class(coarse_dat, "SpatialPolygons"),
     checkmate::check_class(coarse_dat, "sf"),
     checkmate::check_class(coarse_dat, "sfc")
   )
 
   checkmate::assert_class(fine_dat, "RasterLayer")
   checkmate::assert_numeric(d)
+  
 
   if(is_grid & !quiet) {
[…]
