+++ R/create_torus.R
@@ -8,10 +8,19 @@
 #' @param dpad The amount by which to pad the raster (in the same units as the
 #'   raster)
 #'
-#' @return raster. Original raster padded by r cells with torus effect
+#' @return raster. Original raster padded by r cells with torus effect (see
+#'   Details)
 #'
 #' @keywords torus, raster
 #'
+#' @details A torus is an infinite surface where the top joins the bottom, and
+#'   the left side meets the right side. See https://en.wikipedia.org/wiki/Torus
+#'   for a full mathematical description.
+#'
+#'   In this function, the torus effect is achieved by adding the specified
+#'   number of rows of the top of the raster to the bottom (and vice versa) and
+#'   the specified number of rows of the right of the raster to the left (and
+#'   vice versa)
 #' @examples
 #' data(cat_ls)
 #' d <- create_torus(dat = cat_ls, dpad = 5)
