### 2019-05-31 0314d1e
add in last tests, update documentation based on R CMD CHECK
