+++ ropsci-review.Rmd
@@ -1,308 +1,0 @@
----
-title: "Review response"
-author: "Laura J. Graham"
-date: "31/05/2019"
-output: html_document
----
-
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE)
-```
-
-Firstly, thanks to @mbjoseph and @johnbaums for really thorough and helpful reviews. The package now passes goodpractice checks and spelling mistakes have been corrected. 
-
-As a general note, I've added additional functionality to the package - the coarse resolution data (to which the fine resolution data is being aggregated) no longer needs to be gridded. 
-
-First responses to @mbjoseph review: 
-
-# README
-
-> A description in the README that outlines the similarities and differences to raster::focal(), and how the moving window functionality relates to spatial convolution would be helpful. It took me a while to wrap my head around what the package did based on the README, but a more explicit connection to other operations that users may know already could reduce the spin-up time for new users.
-
-I've added text to the readme which provides a package overview. In particular: 
-
----
-The `winmove` function acts as a convenient wrapper to `raster::focalWeight` and `raster::focal` which takes advantage of optimised functions built into the `grainchanger` package. 
-
----
-
->It could be good to point out the target audience more directly in the README - i.e., some explicit statement that a user could identify with - landscape ecologists, or generally folks working with spatial data at different resolutions, etc.
-
-Statement added to readme: 
-
----
-As landscape ecologists and macroecologists, we often need to aggregate data in order to harmonise datasets. In doing so, we often lose a lot of information about the spatial structure and environmental heterogeneity of data measured at finer resolution. An example of this is when the response data (e.g. species' atlas data) are available at a coarser resolution to the predictor data (e.g. land-use data). We developed this method and R package in order to overcome some of these issues. 
-
----
-
->The three panel figure in the README (mwda_schematic.png) I thought would be helpful to wrap my head around what's happening, but there were a few things that I found confusing:
-
->The terms "response" and "predictor" along with f(x) in the first panel might suggest some kind of model, but these aren't explained elsewhere.
-
->It would be nice to have an explanation for why the number of grid cells increases in the second figure. Was a torus created?
-
->An explanation of what the thick black squares represent would be useful.
-
-Apologies, I just stuck this figure in from a paper without any explanation! I've moved it into a separate vignette Background & Motivation and given it a caption, which hopefully answers all the above points: 
-
----
-The above is a graphical representation of the MWDA method. In calculating the MWDA measure, three aspects of scale are considered. Predictor grain is the characteristic spatial scale of a predictor variable, that is, the resolution of the environmental data; scale‐of‐effect determines the appropriate scale of the relationship between predictor and response, for example, an ecological neighbourhood; response grain is the grain of the unit into which you are predicting, that is, the resolution of the response variable (represented by the black lines). Note that the colour scale is unitless. Yellow cells represent ‘high’ values and dark blue cells ‘low’ values. Panel 1 shows a close up of one of the response grain cells in panel 2, whereas panel 2 shows all response grain cells for the study region. Panel 3 shows the study region after aggregation. From [*Graham et al. 2019*](https://doi.org/10.1111/2041-210X.13177).
-
----
-
-> The show_landscape function applied to a categorical raster might be improved by using a discrete colormap (via discrete = TRUE). I was confused briefly by the current example, where the colorbar implies continuity.
-
-Done
-
-> In the README there seem to be some places where "surface" and "landscape" are used interchangeably. Would using one or the other, (or "raster") generalize the language a bit?
-
-Done - changed all to raster
-
-# Implementation and interface
-
[…]
