### 2019-06-01 dc1b9e7
improved warnings and testing data
