+++ data-raw/create_data.R
@@ -5,15 +5,23 @@ cont_ls <- NLMR::nlm_mpd(65, 65, roughness = 0.9)
 cat_ls <- landscapetools::util_classify(cont_ls, weighting = c(0.25, 0.25, 0.25, 0.25))
 
 # create coarse data
-g_raster <- raster::aggregate(cont_ls, 13)
-raster::values(g_raster) <- 1
+g_sf <- sf::st_bbox(cont_ls) %>% 
+  sf::st_as_sfc() %>% 
+  sf::st_buffer(-5) %>% 
+  sf::st_make_grid(cellsize = 11, square = TRUE) %>% 
+  sf::st_sf()
 
-g_sp <- as(g_raster, "SpatialPolygonsDataFrame")
-g_sf <- sf::st_as_sf(g_sf)
+g_sp <- as(g_sf, "Spatial")
+
+g_raster <- raster::crop(cont_ls, g_sf) %>% 
+  raster::aggregate(11)
+
+raster::values(g_raster) <- 1
 
 poly_sf <- sf::st_bbox(cont_ls) %>% 
   sf::st_as_sfc() %>% 
-  sf::st_make_grid(cellsize = 13, square = FALSE) %>% 
+  sf::st_buffer(-5) %>% 
+  sf::st_make_grid(cellsize = 5, square = FALSE) %>% 
   sf::st_sf()
 
 # data for testing output still the same
@@ -34,9 +42,16 @@ fn <- function(x, na.rm = TRUE) {
   sum(2*x)
 }
 
-nm_agg_user <- nomove_agg(g_sp, cat_ls, fn)
+nm_agg_user <- nomove_agg(g_sp, 
+                          cat_ls, 
+                          fn)
 
-wm_agg_shei <- winmove_agg(g_sf, cat_ls, 20, "rectangle", shei, lc_class = 1:4)
+wm_agg_shei <- winmove_agg(g_sf, 
+                           cat_ls, 
+                           12, 
+                           "rectangle", 
+                           shei, 
+                           lc_class = 1:4)
 
 wm_agg_shdi <- winmove_agg(coarse_dat = g_sf, 
                            fine_dat = cat_ls, 
@@ -51,11 +66,24 @@ wm_agg_range <- winmove_agg(coarse_dat = poly_sf,
                             type = "rectangle", 
                             win_fun = var_range)
 
-wm_agg_mean <- winmove_agg(poly_sf, cont_ls, 4, "rectangle", mean, var, is_grid = FALSE)
+wm_agg_mean <- winmove_agg(poly_sf, 
+                           cont_ls, 
+                           4, 
+                           "rectangle", 
+                           mean, 
+                           var, 
+                           is_grid = FALSE)
 
-wm_shei_dat <- winmove(cat_ls, 5, "rectangle", shei, lc_class = 1:4)
+wm_shei_dat <- winmove(cat_ls, 
+                       5, 
+                       "rectangle", 
+                       shei, 
+                       lc_class = 1:4)
 
-wm_mean_dat <- winmove(cont_ls, 2, "rectangle", mean)
+wm_mean_dat <- winmove(cont_ls, 
+                       2, 
+                       "rectangle", 
+                       mean)
 
 user_fn <- function(x, lc_class, ...) {
   return(sum(x == lc_class))
@@ -66,7 +94,31 @@ torus_5 <- create_torus(cat_ls, 5)
 torus_20 <- create_torus(cont_ls, 20)
 
 # data for examples
-usethis::use_data(cont_ls, cat_ls, g_sf, poly_sf, overwrite = TRUE, compress = "bzip2")
+usethis::use_data(cont_ls, 
+                  cat_ls, 
+                  g_sf, 
+                  poly_sf, 
+                  overwrite = TRUE, 
[…]
