### 2019-06-01 8cefe70
remove purrr dependency
