### 2019-06-02 8ba86da
update to suggest patchwork - required for frontpage of pkgdown website
