+++ DESCRIPTION
@@ -37,8 +37,7 @@ Suggests:
     rmarkdown,
     covr,
     ggplot2,
-    landscapetools,
-    patchwork
+    landscapetools
 RoxygenNote: 6.1.1
 Language: en-GB
 VignetteBuilder: knitr
