### 2019-06-02 e055fcf
update example figure to match colour scheme
