### 2019-06-02 e86d797
include rgdal dependency
