### 2019-06-03 #289 @karthik
Thank you so much Laura for this really comprehensive and exhaustive response to the reviewer's comments! :100:
I've never been such an impressive response here or elsewhere. Kudos.

@mbjoseph and @johnbaums Thanks very much for your time and expertise on this review. When you get a chance could you look over Laura's responses and changes and either check off the final item  on your respective checklists (recommending approval) or raise any additional questions? 🙏
