### 2019-06-11 #289 @johnbaums
Awesome, @laurajanegraham - revisions look great!

I noticed one typo:

> an additional utility function for use with simulated **landsacpes**

Regarding `create_torus()` - I think it's fine to leave it in, but maybe clarify in the docs that it's only really intended to be used with simulated data. Docs currently state "Allows for moving window analysis that avoids edge effects (e.g. on simulated landscapes)", which might lead users to apply it to real world data as well in order to avoid edge effects.
