### 2019-06-11 9a78823
update based on final ropensci review comments
