### 2019-06-11 #289 @laurajanegraham
Thanks @mbjoseph and @johnbaums . I've corrected the typo and made clearer in the docs that `create_torus()` is for simulated landscapes.
