### 2019-06-11 #289 @karthik
One more small typo: https://github.com/laurajanegraham/grainchanger/blob/769074f0ec1dc21985973bf7c085ec6283452170/R/calc_functions.R#L102
