+++ README.Rmd
@@ -36,7 +36,7 @@ For more information on the [background to and motivation for](background.html) 
 
 The primary functions of the `grainchanger` package are those which facilitate moving-window (`winmove_agg`) and direct (`nomove_agg`) data aggregation. These functions aggregate fine-grain data (`fine_dat`) to a coarse-grain (`coarse_dat`) using a function specified by the user (`agg_fun`). The moving-window method takes in an additional function (`win_fun`) which smooths the fine-grain data prior to aggregation. 
 
-The moving-window smoothing function is also available in the package (`winmove`), as well as several [built-in functions](functions.html), and an additional utility function for use with simulated landsacpes (`create_torus`).
+The moving-window smoothing function is also available in the package (`winmove`), as well as several [built-in functions](functions.html), and an additional utility function for use with simulated landscapes (`create_torus`).
 
 The `winmove` function acts as a convenient wrapper to `raster::focalWeight` and `raster::focal` which takes advantage of optimised functions built into the `grainchanger` package. 
 
