### 2019-06-12 #289 @laurajanegraham
Oops. I've run the spellcheck now and fixed this (and a couple of others). Thanks all.
