### 2019-06-12 #289 @karthik
Congrats @laurajanegraham, your submission has been approved! 🎉 Thank you for submitting and @mbjoseph and @johnbaums for thorough and timely reviews. To-dos:

```
[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)

```
- [ ] Fix any links in badges for CI and coverage to point to the ropensci URL. (We'll turn on the services on our end as needed)

Welcome aboard!  🎉
