### 2019-06-12 #289 @karthik
@laurajanegraham I would also encourage you to archive this version of the accepted package on Zenodo and add the DOI to your readme.
