### 2019-06-12 9c300c8
include ropensci footer
