+++ README.md
@@ -181,3 +181,5 @@ repository](https://github.com/laurajanegraham/grainchanger).
   - Please note that the `grainchanger` project is released with a
     [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By contributing
     to this project, you agree to abide by its terms.
+    
+[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
