### 2019-06-12 8ac9711
update cran comments, news and version number
