+++ cran-comments.md
@@ -1,10 +1,8 @@
 ## Test environments
-* local Windows 10 x64 install, R 3.5.1
-* ubuntu 14.04 (on travis-ci), R 3.5.2
-* mac os x 10.12.6 (on travis-ci), R 3.5.2
+* l Windows Server 2012 R2 x64 (build 9600) (on AppVeyor), R 3.6.0
+* ubuntu 14.04 (on travis-ci), R 3.6.0
+* mac os x 10.12.6 (on travis-ci), R 3.6.0
 
 ## R CMD check results
 
-0 errors | 0 warnings | 1 note
-
-* This is a new release.
+0 errors | 0 warnings | 0 notes
