+++ man/grainchanger-package.Rd
@@ -17,8 +17,8 @@ Data aggregation via moving window or direct methods. Aggregate a
 \seealso{
 Useful links:
 \itemize{
-  \item \url{https://github.com/laurajanegraham/grainchanger}
-  \item Report bugs at \url{https://github.com/laurajanegraham/grainchanger/issues}
+  \item \url{https://github.com/ropensci/grainchanger}
+  \item Report bugs at \url{https://github.com/ropensci/grainchanger/issues}
 }
 
 }
