+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: grainchanger
 Title: Moving-Window and Direct Data Aggregation
-Version: 0.1.0.9000
+Version: 0.2.0
 Authors@R: c(person("Laura", "Graham", 
              email = "LauraJaneEGraham@gmail.com", 
              role = c("aut", "cre"),
@@ -42,5 +42,5 @@ Suggests:
 RoxygenNote: 6.1.1
 Language: en-GB
 VignetteBuilder: knitr
-URL: https://github.com/laurajanegraham/grainchanger
-BugReports: https://github.com/laurajanegraham/grainchanger/issues
+URL: https://github.com/ropensci/grainchanger
+BugReports: https://github.com/ropensci/grainchanger/issues
