### 2019-06-12 #289 @laurajanegraham
Great news, thank you! I've done most of these things. I've not been able to get the Zenodo DOI (but I drafted the release from the repo after moving to ropensci, which may be the issue).
