+++ 2019-06-25-grainchanger.Rmd
@@ -1,0 +1,65 @@
+---
+slug: grainchanger
+title: Aggregating spatial data with the grainchanger package
+package_version: 0.1.0
+authors:
+  - Laura Graham
+date: 2019-06-25
+categories: blog
+topicid:
+tags:
+- Software Peer Review
+- R
+- community
+# delete the line below
+# if you have no preferred image
+# for Twitter cards
+twitterImg: img/blog-images/2019-06-25-grainchanger/name-of-image.png
+output: 
+  html_document:
+    keep_md: true
+---
+
+Everywhere in this template (YAML, chunk below), you should change "post-template" to the slug of your post, and "2019-06-04" to the publication date.
+
+This template will generate both an .html and an .md. The .md is what you will submit in your pull request to roweb2.
+
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(fig.path = 'img/blog-images/2019-06-04-post-template/')
+```
+
+If using this template, and generating images, after knitting you'll need to copy the folder 'img/blog-images/2019-06-04-post-template/' to ' roweb2/themes/ropensci/static/img/blog-images'. The paths in the resulting Markdown will already be valid so you won't need to edit the Markdown file.
+
+Introduction including outline of the post.
+
+### First awesome section
+
+I like Hugo[^1]. Yes, that is how you add a footnote.
+
+#### First awesome subsection of the first awesome section
+
+Below is an image, for instance your package logo, or a cute puppy picture. Here's how to use a Hugo shortcode to include it. You must use the exact same syntax as below i.e. with the html preserve comments and percent signs, otherwise the knitting will break the syntax.
+
+<!--html_preserve--> {{% figure src = "/img/blog-images/2019-06-04-post-template/name-of-image.png" width = "200" alt = "this is the alternative text" %}}
+<!--/html_preserve-->
+
+Below is another image, a plot, so you might see how to add alternative text in this case.
+
+```{r chunkname, fig.cap='alternative text'}
+plot(1:10)
+```
+
+#### Second awesome subsection
+
+Here's how to use a Hugo shortcode to embed a tweet. We recommend the use of [Hugo shortcodes](https://gohugo.io/content-management/shortcodes/) to include tweets, Vimeo or Youtube videos, gists, etc. Again, use the exact same syntax as below i.e. with the html preserve comments and percent signs, otherwise the knitting will break the syntax.
+
+<!--html_preserve--> {{% tweet 1138216112808529920 %}}
+<!--/html_preserve-->
+
+### Conclusion
+
+Have fun writing your blog post!
+
+Here's how to add the footnote text for your reference above.
+
+[^1]: Hugo! https://gohugo.io/
